backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6

# Data Validation
pydantic==2.5.0

# Authentication (JWT only, no bcrypt)
PyJWT==2.8.0

# AssemblyAI
assemblyai==0.28.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0